+++ CHANGES.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 ### Fixed
 - Fixed `LuaSyntaxOptions.AcceptInvalidEscapes` not suppressing errors in cases where `LuaSyntaxOptions.{AcceptWhitespaceEscape,AcceptHexEscapesInStrings,AcceptUnicodeEscape}` were `false` by @TheGreatSageEqualToHeaven in https://github.com/LorettaDevs/Loretta/pull/116.
+- Fixed single line comments not getting a line break added after them in `NormalizeWhitespace` by @GGG-KILLER in https://github.com/LorettaDevs/Loretta/pull/118.
 
 ### Removed
 - Removed `LanguageNames.{CSharp,FSharp,VisualBasic}` by @GGG-KILLER in https://github.com/LorettaDevs/Loretta/pull/115.